Crate lzma

This crate provides a simple interface to liblzma. LZMA is more commonly known as XZ or 7zip, (as in, files with the .xz or .7z file extension). LZMA compression is fast and aggressive, compressing better than bzip2. liblzma implements the XZ variant, so it can read and write .xz files/streams.

Two interfaces are provided. LZMAReader/LZMAWriter are generic Readers and Writers that can be composed with other Read/Write interfaces. For example, wrap them around a File and you can write data to a file while compressing it on the fly, or stream in an xz file from disk.

compress/decompress are easy to use functions for simple use cases.

See the LZMAReader and LZMAWriter documentation for further details on that interface. compress and decompress are documented here.

Examples

let test_string = "Like tears in rain".to_string();
let mut compressed = lzma::compress(&test_string.into_bytes(), 6).unwrap();
let decompressed = lzma::decompress(&mut compressed).unwrap();
let decompressed_str = String::from_utf8(decompressed).unwrap();

assert_eq!("Like tears in rain", decompressed_str);

Reexports

pub use reader::LZMAReader;

pub use writer::LZMAWriter;

pub use error::LZMAError;

Modules

error

lzma_sys

reader
This module implements LZMAReader.

writer
This module implements LZMAWriter.

Enums

Direction

Constants

EXTREME_PRESET

Functions

compress
Compress buf and return the result.

decompress
Decompress buf and return the result.