Enum lzma::Error

pub enum Error {
    IO(Error),
    ByteOrder(Error),
    Corrupted,
    InvalidProperties,
    MissingMarker,
    HasMoreData,
    NeedMoreData,
    FinishedWithMarker,
}

Errors for Reader and Writer.

Variants

IO

A downstream IO error.

ByteOrder

A downstream byteorder error.

Corrupted

The stream is corrupted.

InvalidProperties

Invalid model values.

MissingMarker

The EOS marker is missing.

HasMoreData

The stream has more data but the uncompressed size has been reached.

NeedMoreData

The stream has finished but the uncompressed size has not been reached.

FinishedWithMarker

The stream has finished unexpectedly with a marker.