Struct lzma::lzma_stream_wrapper::LzmaStreamWrapper

pub struct LzmaStreamWrapper {
    // some fields omitted
}

Methods

impl LzmaStreamWrapper

fn new() -> LzmaStreamWrapper

fn easy_encoder(&mut self, preset: u32, check: lzma_check) -> Result<(), LzmaError>

fn stream_decoder(&mut self, memlimit: u64, flags: u32) -> Result<(), LzmaError>

fn end(&mut self)

fn code(&mut self, input: &[u8], output: &mut [u8], action: lzma_action) -> LzmaCodeResult

Pointers to input and output are given to liblzma during execution of this function, but they are removed before returning. So that should keep everything safe.